requirements.txt
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
idna==3.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
pyTelegramBotAPI==4.15.2
python-dotenv==1.0.0
requests==2.31.0
service-identity==24.1.0
TelegramBotAPI==0.3.2
urllib3==2.1.0
